requirements.txt
streamlit
PyPDF2
nltk
openai
fpdf
python-dotenv
